LayoutTests/fast/frames/seamless/seamless-designMode-expected.txt

FAIL window.iframe.contentDocument.body.textContent should be PASS. Was FAIL.
